Simplify plugin links handling
