+++ src/Main.php
@@ -14,7 +14,6 @@ if ( ! defined( 'ABSPATH' ) ) {
 use PostNLWooCommerce\Checkout_Blocks\Blocks_Integration;
 use PostNLWooCommerce\Checkout_Blocks\Extend_Block_Core;
 use PostNLWooCommerce\Checkout_Blocks\Extend_Store_Endpoint;
-use PostNLWooCommerce\Admin\Plugin_Links;
 
 /**
  * Class Main
@@ -124,8 +123,7 @@ class Main {
 		add_action( 'block_categories_all', array( $this, 'register_postnl_block_category' ), 10, 2 );
 		add_filter( 'woocommerce_shipping_methods', array( $this, 'add_shipping_method' ) );
 
-		$links = new Plugin_Links(POSTNL_WC_PLUGIN_BASENAME);
-		$links->register();
+		$this->register_plugin_links();
 
 	}
 
@@ -465,6 +463,66 @@ class Main {
 				),
 			)
 		);
+	}
+
+	/**
+	 * Register plugin action/meta links (admin only).
+	 *
+	 * @return void
+	 */
+	public function register_plugin_links() {
+		if ( ! is_admin() ) {
+			return;
+		}
+
+		add_filter( 'plugin_row_meta', array( $this, 'add_row_meta' ), 10, 2 );
+		add_filter( 'plugin_action_links_' . POSTNL_WC_PLUGIN_BASENAME, array( $this, 'add_action_links' ), 10, 1 );
+	}
+
+	/**
+	 * Add row meta links.
+	 *
+	 * @param string[] $links Existing links.
+	 * @param string $file Plugin file name.
+	 *
+	 * @return string[]
+	 */
+	public function add_row_meta( array $links, string $file ): array {
+		if ( $file === POSTNL_WC_PLUGIN_BASENAME ) {
+			$links[] = sprintf(
+				'<a href="%s" target="_blank" rel="noopener">%s</a>',
+				esc_url( 'https://wordpress.org/support/plugin/woo-postnl/reviews/#new-post' ),
+				esc_html__( 'Leave a review', 'postnl-for-woocommerce' )
+			);
+		}
+
+		return $links;
+	}
+
+	/**
+	 * Add action links.
+	 *
+	 * @param string[] $links Existing links.
+	 *
+	 * @return string[]
+	 */
+	public function add_action_links( array $links ): array {
+		array_unshift(
+			$links,
+			sprintf(
+				'<a href="%s">%s</a>',
+				esc_url( admin_url( 'admin.php?page=wc-settings&tab=shipping&section=postnl' ) ),
+				esc_html__( 'Settings', 'postnl-for-woocommerce' )
+			)
+		);
+
+		$links[] = sprintf(
+			'<a href="%s" target="_blank" rel="noopener">%s</a>',
+			esc_url( 'https://wordpress.org/support/plugin/woo-postnl/reviews/#new-post' ),
+			esc_html__( 'Leave a review', 'postnl-for-woocommerce' )
+		);
+
+		return $links;
 	}
 
 	/**
